requirements.txt
fuzzywuzzy
python-dotenv